+++ src/poptracker.cpp
@@ -27,6 +27,9 @@ using nlohmann::json;
 using Ui::Dlg;
 
 
+const Version PopTracker::VERSION = Version(APP_VERSION_TUPLE);
+
+
 enum HotkeyID {
     HOTKEY_TOGGLE_VISIBILITY = 1,
     HOTKEY_RELOAD,
@@ -100,7 +103,11 @@ PopTracker::PopTracker(int argc, char** argv, bool cli, const json& args)
         }
 
 #ifndef WITHOUT_UPDATE_CHECK
-        if (_config.find("check_for_updates") == _config.end()) {
+        if (!Dlg::hasGUI()) {
+            printf("Dialog helper not available! "
+                    "Please install which and zenity, kdialog, matedialog, qarma, xdialog or python + tkinter.\n");
+        }
+        else if (_config.find("check_for_updates") == _config.end()) {
             auto res = Dlg::MsgBox("PopTracker",
                     "Check for PopTracker updates on start?",
                     Dlg::Buttons::YesNo, Dlg::Icon::Question);
@@ -172,15 +179,15 @@ PopTracker::PopTracker(int argc, char** argv, bool cli, const json& args)
     _homePackDir = os_pathcat(homePath, "PopTracker", "packs");
     _appPackDir = os_pathcat(appPath, "packs");
 
-    if (!homePath.empty() && homePath != "." && homePath != cwdPath) {
+    if (!homePath.empty()) {
         Pack::addSearchPath(_homePackDir); // default user packs
         if (!_isPortable)
-            Assets::addSearchPath(os_pathcat(homePath,"PopTracker","assets")); // default user overrides
+            Assets::addSearchPath(os_pathcat(homePath, "PopTracker", "assets")); // default user overrides
     }
-    if (!documentsPath.empty() && documentsPath != "." && documentsPath != cwdPath) {
-        Pack::addSearchPath(os_pathcat(documentsPath,"PopTracker","packs")); // alternative user packs
+    if (!documentsPath.empty() && documentsPath != ".") {
+        Pack::addSearchPath(os_pathcat(documentsPath, "PopTracker", "packs")); // alternative user packs
         if (_config.value<bool>("add_emo_packs", false)) {
-            Pack::addSearchPath(os_pathcat(documentsPath,"EmoTracker","packs")); // "old" packs
+            Pack::addSearchPath(os_pathcat(documentsPath, "EmoTracker", "packs")); // "old" packs
         }
     }
     if (!appPath.empty() && appPath != "." && appPath != cwdPath) {
@@ -811,6 +818,12 @@ bool PopTracker::ListPacks(PackManager::confirmation_callback confirm, bool inst
         printf("\n");
     }
 
+    printf("Search paths:\n");
+    for (const auto& path: Pack::getSearchPaths()) {
+        printf("%s\n", path.c_str());
+    }
+    printf("\n");
+
     printf("Installed packs:\n");
     auto installed = Pack::ListAvailable();
     if (installed.empty()) {
@@ -918,10 +931,17 @@ bool PopTracker::loadTracker(const std::string& pack, const std::string& variant
     
     printf("Loading Pack...\n");
     _pack = new Pack(pack);
+
+    bool targetLatest = !(_pack->getTargetPopTrackerVersion() > Version{});
+    const Version& targetVersion = targetLatest ? VERSION : _pack->getTargetPopTrackerVersion();
+    printf("Target PopTracker version: %s%s\n",
+            targetVersion.to_string().c_str(),
+            targetLatest ? " (undefined/latest)" : "");
+
     printf("Selecting Variant...\n");
     _pack->setVariant(variant);
 
-    if (_pack->getMinPopTrackerVersion() > Version(VERSION_STRING)) {
+    if (_pack->getMinPopTrackerVersion() > VERSION) {
         Dlg::MsgBox("Error", "Pack requires PopTracker " + _pack->getMinPopTrackerVersion().to_string() + " or newer. "
                     "You have " + VERSION_STRING + ".");
         fprintf(stderr, "Pack requires a newer version of PopTracker!\n");
@@ -1076,7 +1096,7 @@ bool PopTracker::loadTracker(const std::string& pack, const std::string& variant
         _tracker->AddLocations("locations.json");
 
     // run pack's init script
-    printf("Running scripts/init.lua\n");
+    printf("Running init...\n");
     bool res = _scriptHost->LoadScript("scripts/init.lua");
     // save reset-state
     StateManager::saveState(_tracker, _scriptHost, _win->getHints(), json::value_t::null, false, "reset");
@@ -1255,7 +1275,7 @@ void PopTracker::updateAvailable(const std::string& version, const std::string& 
 bool PopTracker::isNewer(const Version& v)
 {
     // returns true if v is newer than current PopTracker
-    return v>Version(VERSION_STRING);
+    return v > VERSION;
 }
 
 bool PopTracker::saveConfig()